+++ packages/orchestration/src/examples/unbond.contract.js
@@ -1,8 +1,8 @@
 import { M } from '@endo/patterns';
 import { withOrchestration } from '../utils/start-helper.js';
+import * as flows from './unbond.flows.js';
 
 /**
- * @import {Orchestrator, OrchestrationFlow} from '../types.js'
  * @import {TimerService} from '@agoric/time';
  * @import {LocalChain} from '@agoric/vats/src/localchain.js';
  * @import {NameHub} from '@agoric/vats';
@@ -11,37 +11,6 @@ import { withOrchestration } from '../utils/start-helper.js';
  * @import {CosmosInterchainService} from '../exos/cosmos-interchain-service.js';
  * @import {OrchestrationTools} from '../utils/start-helper.js';
  */
-
-/**
- * @satisfies {OrchestrationFlow}
- * @param {Orchestrator} orch
- * @param {object} ctx
- * @param {ZCF} ctx.zcf
- * @param {ZCFSeat} _seat
- * @param {undefined} _offerArgs
- */
-const unbondAndLiquidStakeFn = async (orch, { zcf }, _seat, _offerArgs) => {
-  console.log('zcf within the membrane', zcf);
-  // We would actually alreaady have the account from the orchestrator
-  // ??? could these be passed in? It would reduce the size of this handler,
-  // keeping it focused on long-running operations.
-  const omni = await orch.getChain('omniflixhub');
-  const omniAccount = await omni.makeAccount();
-
-  // TODO implement these
-  // const delegations = await celestiaAccount.getDelegations();
-  // // wait for the undelegations to be complete (may take weeks)
-  // await celestiaAccount.undelegate(delegations);
-  // ??? should this be synchronous? depends on how names are resolved.
-  const stride = await orch.getChain('stride');
-  const strideAccount = await stride.makeAccount();
-
-  // TODO the `TIA` string actually needs to be the Brand from AgoricNames
-  // const tiaAmt = await celestiaAccount.getBalance('TIA');
-  // await celestiaAccount.transfer(tiaAmt, strideAccount.getAddress());
-  // await strideAccount.liquidStake(tiaAmt);
-  console.log(omniAccount, strideAccount);
-};
 
 /**
  * Orchestration contract to be wrapped by withOrchestration for Zoe
@@ -58,12 +27,8 @@ const unbondAndLiquidStakeFn = async (orch, { zcf }, _seat, _offerArgs) => {
  * @param {Zone} zone
  * @param {OrchestrationTools} tools
  */
-const contract = async (zcf, privateArgs, zone, { orchestrate }) => {
-  const unbondAndLiquidStake = orchestrate(
-    'LSTTia',
-    { zcf },
-    unbondAndLiquidStakeFn,
-  );
+const contract = async (zcf, privateArgs, zone, { orchestrateAll }) => {
+  const { unbondAndLiquidStake } = orchestrateAll(flows, { zcf });
 
   const publicFacet = zone.exo('publicFacet', undefined, {
     makeUnbondAndLiquidStakeInvitation() {
